chore: adding support for javascript
